Add convenience classes to wrap Runnable instances into PlugIns
With the new concept of input/output parameters being encoded as annotated
fields, all that is left of the generic plugin interface is a public run()
method, i.e. the Runnable interface.

Provide ways (time will tell which ones are the most useful) to wrap
Runnables as plugins:

	PlugIn plugin = new RunnableAdapter(new MyPlugIn());

or

	public class Wrapped_PlugIn extends RunnableAdapter {
		public Wrapped_PlugIn() {
			super(new MyPlugIn());
		}
	}

or

	public class Wrapped_PlugIn extends PlugInWrapper {
		public Wrapped_PlugIn() {
			super("MyPlugIn");
		}
	}

or in a plugins.config:

	Bla>Blub, "Blizz", fiji.plugin.PlugInWrapper("MyPlugIn")

Unfortunately, it is not possible to do this with generics
(Bla>Blub, "Blizz", fiji.plugin.PlugInWrapper<MyPlugIn>), because
Java generics work by erasure.  Therefore, a PlugInWrapper<T>
cannot instantiate T.

But a possible option is to teach ImageJA to fall back to using the
PlugInWrapper if it finds that class and the plugin is an instance
of a Runnable.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>


git-svn-id: http://code.imagej.net/svn/imagej/trunk@1150 be23795f-a60e-40d4-ba19-055798872e57