Initial proof-of-concept for imglib + ImageJ.
Replaces ByteProcessor's "byte[] pixels" with an imglib Image<ByteType>
using a byte array container. Has lots of hacks to grab pixels directly
from the container, etc., but for legacy ImageProcessor support this approach
may be good enough. In the long term, perhaps we can migrate away from the
ImageProcessor subclasses toward an "ImgLibProcessor" or some such that is
backed by an arbitrary imglib Image object.

There is still much work to be done modularizing the ImageProcessor
functionality, but that is a separate issue.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@1173 be23795f-a60e-40d4-ba19-055798872e57