Proof-of-concept version of ImgLibProcessor.
This class allows an ImagePlus to be backed by an imglib image
(i.e., an Image<? extends ComplexType> object). This version
implements only enough methods to display an image stack onscreen.

There is currently case logic in static methods for extracting primitive
arrays of various types for a given image plane of an Image. The need
for this case logic is unfortunate, but we could not immediately see a
better way. In cases where the Image is not a known Type, we return a
double[] array by calling ComplexType.getRealDouble(), which will not
work elegantly for complex data.

NB: The getPixels() method always creates a copy of the plane, which
if modified by reference will not update the linked ImagePlus unless
setPixels is later called (and setPixels is still unimplemented ;-).


This used to be revision r1258.