Build using Java 1.6, since imglib requires 1.6.
Without this change, there were many class version errors when
executing imglib-related functionality.

Unfortunately, there is a still an issue with imagej where the special
tools.jar dependency causes m2eclipse to create an Eclipse project
linked to J2SE1.5 rather than 1.6, which results in the same problem.
Further investigation is required as to the best way to address this.

There is also a question of whether it is acceptable to require 1.6
for ImageJ at runtime. The main motivation for using 1.5 is to
support OS X 10.4, which does not have a 1.6 JRE, but recently Apple
has announced that they are discontinuing their JVM anyway. Perhaps
OpenJDK on OS X will improve and 1.6 will be possible on OS X 10.4?


This used to be revision r1597.