Progress on legacy ImageJ bridge mechanism.
Conversion of ImagePlus objects to Datasets is now ostensibly
functional. For the moment, only grayscale single-plane images work, but
the groundwork is in place to support higher dimensional images.

The bridge uses Javassist to override ij.gui.ImageWindow.show(), so that
rather than actually displaying the ImageWindow on screen, it instead
registers the corresponding ImagePlus with the LegacyManager, which
keeps a table of ImagePlus objects and their corresponding Datasets.

Whenever an ImagePlus is shown onscreen, a dummy plugin is triggered
that produces the linked Dataset as output, signaling any relevant
postprocessors (e.g., ij2-gui's imagej.gui.DisplayPostprocessor) that
the Dataset has become available. Sadly, this is a hacky mechanism for
displaying Datasets, and requires further thought, but works for now.

The end result is that if a legacy ImageJ plugin (e.g., File > Open)
creates and shows an ImagePlus, the ij2-gui interface instead displays
the corresponding Dataset.

Hooks are in place to go the other way (i.e., execute legacy ImageJ
plugins on Datasets, treating them as ImagePlus objects), but it is not
yet fully functional. The main obstacle is that we do not yet have a
mechanism to programmatically identify the currently active Dataset.


This used to be revision r2175.