Add another level of hierarchy to the ImageJ projects.
This includes a division of projects between "core" and "non-core"
as well as a separate subtree for Envisaje, and another for
unused/obsolete projects.

It is now more convenient to build just the "core" ImageJ2 by
navigating to the core subtree and building from there.


This used to be revision r2202.