Begin reconciliation of ImageJ2 plugins and workflows.
Terminology:
  * A "module" is a processing node with typed inputs and outputs
  * A "workflow" is a graph of modules, which is itself a module
  * A "tool" is a class that invokes modules in response to user input
  * A "plugin" is type of module defined as Java code that implements
    the BasePlugin interface (usually a subclass thereof):
    * Plugins are labeled with annotations and indexed as "@Plugin"
      using SezPoz
    * Plugins can define their parameters with annotations
    * Plugins can use annotations to suggest how they should be exposed
      in a user interface
    * Types of plugins are differentiated by the "type" annotation field

Because the plugin type hierarchy consists solely of interfaces, in
order for plugins to implement the module interface hierarchy as well,
there must be adapter classes providing the functionality. The plugin
framework already essentially provided it in the form of the
PluginHandler class, but it did not actually implement the requisite
module interfaces; the new PluginModule and PluginModuleInfo classes
now do so.

There is still work left to do, but these changes constitute the bulk
of the integration between core/ij-plugin and extra/workflow.



git-svn-id: http://code.imagej.net/svn/imagej/trunk@2217 be23795f-a60e-40d4-ba19-055798872e57