Initial cut at GUI-agnostic user input events.
Also added some initial tool events. The goal is to tie tools and
displays together through publication and subscription of these events.

Next step is to add a tool manager class that keeps track of an active
tool, subscribes to the user input events, and passes them along to
the currently active tool. The ITool interface will change
significantly as a consequence.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@2297 be23795f-a60e-40d4-ba19-055798872e57