Fix preferences key uniqueness.
The problem is that java.util.prefs.Preferences uses the package name
of the given class, but not the class name itself, when storing
preferences. So we must incorporate the given class's simple name
into the key itself. E.g., for the inputFile parameter of
imagej.core.plugins.OpenImage, we use OpenImage.inputFile for the key
by default.

Otherwise, two plugins within the same package would overwrite each
others' persisted values for parameters with the same name.


This used to be revision r2366.