Initial cut at updating dataset's existing display.
Also made some miscellaneous tweaks and change to the display logic.

The way it works right now is suboptimal; the display postprocessor
should not be responsible for manually updating displays, but rather
the displays should listen for DatasetChangedEvents to update
themselves.

Lastly, there is a problem currently where plugins are outputting
a new dataset, rather than the same object as was inputted.


This used to be revision r2531.