Initial, untested update of ROI logic to JHotDraw 7.
The translation was straightforward except for the MDI application in
ExampleGUI. I adapted code from org.jhotdraw.samples.draw.DrawingPanel
to populate the JToolBar using the discovered JHotDraw adapters, but
have not yet determined how best to add that toolbar to the application.
There is also likely an issue with the associated action, defined in a
ResourceBundleUtil table.

This commit also has some style cleanup and refactoring.


This used to be revision r2667.