Fix problem where Datasets were never deleted.
It works by performing reference counting. Each DatasetView increments
the dataset's reference counter, and when the display is closed it
disposes all views, which decrement the counter. When the counter
reaches zero, the dataset is "deleted" (i.e., DatasetDeletedEvent is
published, and ObjectManager removes the dataset from its hoard).


This used to be revision r2783.