Initial API changes for handling overlays. WIP.
This approach is ultimately untenable in certain ways. We are planning
to rework the display logic to use JHotDraw as the primary rendering
layer in the Swing UI (for both overlays and datasets). For the moment,
this commit is an initial effort to render overlays in the same display
as a dataset, using JHotDraw. Several more substantial changes to the
API will be needed before this functionality works properly.


This used to be revision r2821.