Generalize line color plugin to overlay properties.
It may be fine to have a single overlay properties plugin that allows
people to change various properties. The only downside is that right
now you cannot "opt out" of changing a subset of the properties, so if
you want to e.g. change only the line colors, but leave mixed
selection line widths alone, you can't.


This used to be revision r2909.