Add ability to get dataset's associated overlays.
This is a hacky method that should probably go away in the future.
It would make more sense for the legacy layer to map one ImagePlus per
IJ2 Display (and also one ImagePlus per orphaned Dataset). In this
way, each Display's linked Overlays can be associated with only its
relevant ImagePlus. Right now, to maintain the ImagePlus <-> Dataset
mapping, we are instead unioning all Overlays of all Displays
containing a particular Dataset, which is highly unintuitive.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@2951 be23795f-a60e-40d4-ba19-055798872e57