Initial overlay syncing between IJ1 and IJ2.
Works halfway at the moment, in that Overlays are translated into
ImagePlus ROIs, but not vice versa. So you can now do things like
create a rectangular selection, run a legacy IJ1 plugin such as
Despeckle and see that only the selected region was affected.
But other plugins such as Edit > Selection > Enlarge will not be
reflected in IJ2 data structures yet.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@2952 be23795f-a60e-40d4-ba19-055798872e57