Make input resolution instance-specific.
The setResolved and isResolved methods were previously in the
PluginModuleItem class, which is the wrong place, because ModuleItem
is intended to contain metadata about the module, but not state for
instances of the module. Instead, the methods have been moved to the
Module interface where they belong.

Also, this changeset updates the ActiveDatasetPreprocessor (renamed to
ActiveDisplayPreprocessor) to populate both single Dataset inputs and
single Display inputs with whatever is currently active.


This used to be revision r3045.