Mechanism for handling when no datasets are open.
This initial approach allows preprocessors to register a reason for
plugin cancelation, which the plugin runner emits as a StatusEvent
warning, which the UI intercepts and displays to the user in a dialog.

This commit also fleshes out various javadoc comments, and has minor
code cleanup in a few places.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@3080 be23795f-a60e-40d4-ba19-055798872e57