Eliminate static "isInsideIJ1" logic.
It would be better to avoid globally static flags like this. ImageJ1 is
filled with them and it can become a mess very quickly. At worst, we can
add such state flags to the LegacyManager instance, and refer to that
instead, but hopefully we can avoid state flags completely.

This used to be revision r3214.