Added DisplaySelectedEvent, modified DisplayManager, SwingImageDisplay...
DisplayManager modified so that upon receiving a WinActivatedEvent, is sets the active display and publishes a DisplaySelectedEvent.

In SwingImageDisplay, added:
List<EventSubscriber<?>> subscribers;
and subscribeToEvents() method is invoked on construction.

imagej.ui.swing.plugins WatchDisplays and WatchOverlays use DisplaySelectedEvent to correctly update their status.
(See ticket #633)

This used to be revision r3260.