Rename DisplaySelectedEvent to DisplayActivatedEvent.
The rationale is that "selected" is a term used by the DisplayView
events to indicate that that view has been selected or deselected.
Multiple selections are allowed. Conversely, in the case of the active
display, exactly one is allowed to be active at one time. We also refer
to it in the DisplayManager API as "setActiveDisplay" rather than
"setSelectedDisplay" so we should keep our names consistent.

This used to be revision r3263.