Change terminology from 'Manager' to 'Service'.
The rationale is that 'Service' is a more commonly used term to describe
such modular areas of functionality. Further, the term 'Manager' was
used in IJ1, from an end-user perspective, to mean a GUI component for
managing collections of objects (e.g., ROI Manager). Now that we have
our own Overlay Manager, it will become increasingly confusing to
continue to refer to the core ImageJ services as "managers" too.

This used to be revision r3333.