More refactoring of module framework:
  * Add AbstractModuleInfo implementation.
  * Rename UserInterfaceDetails to UIDetails.
  * Make UIDetails extend BasicDetails, for simplicity.
  * Split out core UIDetails implementation from IndexItemInfo.
  * Reuse AbstractUIDetails implementation in AbstractModuleInfo.

Note that PluginModuleInfo still extends PluginInfo rather than
AbstractModuleInfo, and they replicate a small amount of code for the
ModuleInfo method implementations. Currently, AbstractModuleInfo is not
used by anything in the core ImageJ implementation, but it will be
useful later when implementing custom modules.

This used to be revision r3345.