Merge event and object with core, for simplicity.
The rationale is that the event framework and object service are both
very central to the ImageJ paradigm, and do not add much overhead to the
JAR size. This does introduce one additional dependency (eventbus) to
ij-core, but ij-core already has some dependencies, so one more is not
particularly more complex than before for deployment.

Many subprojects needed to include all three of core, event and object
(or at least two of the three) anyway, so hopefully the merge makes the
ImageJ2 design easier to understand.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@3368 be23795f-a60e-40d4-ba19-055798872e57