Combine module and plugin to ext, for simplicity.
At first I was reluctant to do this, because plugin depends on sezpoz,
but module depends on core which also depends on sezpoz in any case, so
it is not really a new dependency. The imagej.ext.plugin classes do not
add too much overhead to the JAR; it can still be used for "pure"
module-driven implementations with no use of SezPoz, if desired.

This used to be revision r3369.