Add generation of single ImageJ JAR bundle.
This was tricky because combining multiple JARs together overwrites each
of their SezPoz annotations. To fix, we keep the combined annotations
files in the assembly's resources directory. These combined annotations
were generated by a special program,
src/test/java/imagej/util/CombineAnnotations.java,
which can be run from an IDE to regenerate them whenever they change.

However, this solution is far from optimal as it requires the combiner
to be rerun every time a new plugin or other dynamically discovered
class is added to the core distribution. It would be better if these
files were not stored in the VCS, but were instead generated on the fly
as part of the assembly execution.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@3674 be23795f-a60e-40d4-ba19-055798872e57