Use TypeChanger for the plugin type.
This is more specific than ImageJPlugin and allows to query
PluginService.getRunnablePluginsOfType(TypeChanger.class) to access only
these specific plugins.

Now that the PluginService's index is hierarchical and this works, we
should probably use more specific types whenever multiple plugins share
a common structure like this.

This used to be revision r3695.