Use clamped value instead of throwing exception.
This partially fixes plugins where the initial value falls outside the
min/max range. However, the initial value is still ultimately set in the
spinner (at least in the Swing UI). The spinner arrows then do not work,
as they always do nothing when confronted with an invalid starting
value.

We could fix it so that the value is clamped elsewhere (i.e.,
investigate further to discover where the model value is being set
outside the min/max range, and clamp there), but it does not seem worth
the effort right now, when plugins should be declaring proper initial
values in any case. However, eventually this problem may cause trouble
for plugin developers in the community, so we may want to fix later.

This used to be revision r3848.