Add the ScriptFinder
The purpose of this class is -- similar to the legacy plugin finder -- to
look through ImageJ.app/plugins/ and find all the scripts. Those scripts
will be added to the Plugins menu (or submenu, corresponding to the
subdirectory where the scripts were found). All the hard work is handed
off to the ScriptPlugin class.

Like in Fiji, the directory ImageJ.app/plugins/Scripts/ is special in that
the scripts will be pulled up into the top-level menu structure, e.g.
ImageJ.app/plugins/Scripts/File/Fetch.js will be available as File>Fetch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>