Switch "newOutputWindow" API back to OutputWindow.
Output windows are more broadly applicable than text displays; anything
can create a new output window and append text to it, whereas a text
display is specifically a Display implementation for strings, the
default implementation of which will be backed by an output window.

This used to be revision r3877.