Allow module items to be both inputs and outputs.
Analogously, this loosens the restriction on plugin parameters as well.
To accommodate this, we have a new enum called ItemIO with three
choices: input, output, or both. The ModuleItem API now provides new
methods getIOType, isInput and isOutput for querying whether a module
item is an input or an output. Lastly, these changes fix a bug in
ModuleItem.getValue, which previously assumed the item was an input.

This used to be revision r3893.