Remove output event.
The different logging levels are cool, but the issue is that output
windows are not singletons anymore. So if e.g. multiple
SwingOutputWindows are created and then an OutputEvent is published, all
output windows will be updated with the output.

Since nothing is using the OutputEvent currently, it is easier to remove
it for now.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@3908 be23795f-a60e-40d4-ba19-055798872e57