When building a path to find native libraries, use "gpp" rather than "g++" as the linker for Linux versions.
The LOCI maven repository has "slim-curve-1.0-[date/time stamp]-amd64-Linux-gpp-shared.nar".  We don't really care what linker was used but it's important that the nar file name and the plugin jar directory that the native library contained in the nar file winds up in have consistent names.


This also turns on IJ.log logging.

This used to be revision r3919.