Prohibit the status bar from grabbing the focus all the time
There was a hack in place to keep the progress bar's blue color on MacOSX
by grabbig the focus all the time. Unfortunately, on other operating
systems, this leads to ImageJ constantly grabbing the focus even when the
user switched to another application altogether. So let's get rid of this
hack.

Unfortunately, I did not find any way to keep the progress bar in blue
without this hack. But then, turning the progress bar into a gray thing
when the focus is somewhere else is a part of the MacOSX Look & Feel.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r3933.