Remove ij.Prefs.set injection hook.
This change prevents an infinite loop where the options synchronizer
called ij.Prefs.set which called the options synchronizer, etc.

For the most part, when an IJ1 option changes, it is due to a legacy
plugin execution, so the LegacyPlugin calling the options synchronizer
post-execution is sufficient. However, for legacy plugins that spawn a
new thread or open a non-modal dialog, there is a remaining issue in
that the synchronization will happen *before* the plugin's logic
finishes executing in the other thread. This is the situation with the
legacy IJ1 options plugins, because they bring up non-modal dialogs.

In the case of the legacy options plugins, we may not need to fix the
issue, since we have blacklisted those plugins already. The issue in
general merits further thought, though.

This used to be revision r3939.