Use shorter note about load() in constructors.
Despite my initial trepidation, this solution works well. The only other
way I could think to address the problem is to add a load() call in
OptionService anywhere it creates a new OptionsPlugin. But that approach
could be insufficient if any code instantiates an options plugin outside
the options service. So let's stick with the constructor-based approach!

git-svn-id: http://code.imagej.net/svn/imagej/trunk@3943 be23795f-a60e-40d4-ba19-055798872e57