Move options framework and plugins to ij-options.
This is similar to the ij-io project in that it includes some core
plugins (in imagej.options.plugins), but also the relevant framework
classes. This is a cleaner organization for ij-legacy to depend on it,
since currently nothing depends on the ij-plugins-* projects (other than
ui/imagej at runtime), and it would be better to keep it that way.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@3947 be23795f-a60e-40d4-ba19-055798872e57