Move options framework and plugins to ij-options.
This is similar to the ij-io project in that it includes some core
plugins (in imagej.options.plugins), but also the relevant framework
classes. This is a cleaner organization for ij-legacy to depend on it,
since currently nothing depends on the ij-plugins-* projects (other than
ui/imagej at runtime), and it would be better to keep it that way.

This used to be revision r3947.