Fix zoom center error
This fixes the issue described in ticket 628 (zooming would not center at
the mouse). It was due to a sign error (the offset needed to be added,
rather than subtracted, because it is in panel coordinates, not in image
coordinates).

While at it, avoid granting external classes access to the private
'offset' field's content by returning a copy instead of the object itself
(which could be subsequently be modified by the caller).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@3961 be23795f-a60e-40d4-ba19-055798872e57