Fix ticket 713
Zoom to selection was broken because the JScrollPane's viewport was not
given a chance to update its dimension, and therefore the request to set
the origin correctly was denied, defaulting to (0,0).

Since changing the zoom also resets the offset, we have to remember the
pan offset and tell CanvasHelper to reset it after validating the
JScrollPane's viewport.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r3963.