Fix up Swing status bar to be more like IJ1.
The UI now uses a JLabel with lowered bevel border, with a progress bar
on the right that only appears when there is progress to report.

This update also uses an explicit event service rather than the static
Events class for working with status events.

This used to be revision r3967.