Move spatial methods into LabeledSpace interface.
These methods are common to all LabeledSpace implementations (i.e., both
Data and ImageDisplay), so no reason not to unify their declarations.

This also unifies the redundant methods "Axis[] getAxes()" and
"List<Axis> getAxes()" in favor of the former only.

There is still a lot of unnecessary array allocation in the core
implementations -- to be fixed later.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4066 be23795f-a60e-40d4-ba19-055798872e57