Fix Set Zoom plugin.
The Eclipse cleanup rules add "final" to any variable that cannot be
changed (e.g., private instance fields that are assigned exactly once).
To avoid this in plugins, you can use Eclipse's "Generate Getters and
Setters" source command to quickly create getters and setters for all
input parameters. And we need these methods anyway, as we must provide
public API on a per-plugin basis for assigning input parameter values.

This used to be revision r4072.