Fix compile errors in IJ1 unit tests.
The Calibration.isSameAs method still needs to be readded (as
CalibrationTools.isSame(Calibration, Calibration)). For now it is coded
to always pass.

It should now be possible to begin addressing failing tests and report
any regression bugs to Wayne. Once they are resolved, we can add the
ij1-tests build to the Hudkins.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4092 be23795f-a60e-40d4-ba19-055798872e57