Progress on AutoContrast plugin.
I added a few notes to key areas of code where autoscaling happens, that
we need to revisit how autoscaling is done in general.

There is one remaining problem with AutoContrast with non-8-bit data.
The issue is that there is a conflation between bin numbers and pixel
values. So it (I believe) correctly computes the histogram bin range
across which to autoscale, but then sets the min/max pixel values to
those bin values. For 8-bit data, this works, because the 0-255 range is
congruent. For >8-bit data, this results in weird pixel values leading
to a nearly-all-white image display.

The fix is simply to keep track of the histogram's min and max and
translate the bin values back into pixel intensity values, before
assigning the min & max.

This used to be revision r4106.