Track modifier flags for all input events.
KyEvent and MsEvent now extend a common InputEvent, which has an
InputModifiers object that indicates which modifiers are active
(including both keyboard modifier keys and mouse buttons).

On the AWT side, the keyboard and mouse event dispatchers extend a
common class with helper methods for creating InputModifiers objects.

With this change, it should now be easy to detect input combinations
such as Ctrl + left mouse button click.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4144 be23795f-a60e-40d4-ba19-055798872e57