Eliminate explicit os.name case logic.
The new approach looks for a com.apple.eawt class, and if found,
overrides IJ1's MacAdapter behavior. Conversely, if the class is not
found, no override is necessary (or even possible).

This logic is necessary because IJ2 installs its own Mac OS X
application hooks. When IJ1 attempts to install its application handler,
it conflicts with IJ2's already installed handler, throwing an
exception. Since the handler's purpose is to support Mac-specific menu
items (specifically About, Preferences and Quit in the application
menu), and since we want to run IJ1 invisibly anyway, we silently
prevent IJ1 from installing its handler in favor of the IJ2 handler.

This used to be revision r4220.