KyEvent: fix accelerator string building
When we typed a modifier, the character is negative, so let's use that as
a tell-tale that we should not add the character to the accelerator
string.

Further, when a combination like Control+Q was typed, the character is not
'Q', but a non-ASCII character. So let's use KeyCode's toString() method
to get at the key.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4222.