Rename imagej subproject to ij-app.
This change avoids a bug in Eclipse that occurs when checking out a
working copy to a folder with the same name as one of its subprojects.
Because there was a project called "imagej" you would encounter problems
if your working copy was stored in workspace/imagej.

The bug is that Eclipse insists on naming the toplevel project to match
the folder name beneath workspace (in this case, "imagej") even though
the toplevel POM says otherwise (in this case, "imagej-pom-base").

Since Eclipse is incapable of having two projects in the same workspace
with the same name, the actual "imagej" project cannot be imported,
preventing (among other things) imagej/Main.java from being executed.

By naming the project "ij-app" instead, we avoid the name clash while
also ensuring it is first on the list of projects alphabetically.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@4230 be23795f-a60e-40d4-ba19-055798872e57