Dispatch app frame events only to certain tools.
Most tools want to ignore events from the main app frame. But there is
now a flag, activeInAppFrame, that a tool can set if it wants to receive
these events. The display from these events will be null, but the tool
can ask for the active image display if it wants to function that way.
(Or it can behave entirely differently if needed!)

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4286 be23795f-a60e-40d4-ba19-055798872e57