Fix bug with non-control accelerators.
Plugins like AxisPositionForward and AxisPositionBackward use
accelerators that do not include the control modifier. In their case,
they are mapped to "shift PERIOD" and "shift COMMA" respectively. But
the AcceleratorHandler was always tacking on "control" when the user
pressed "shift PERIOD" so the module accelerator lookup always failed.

The fix is to first look up the exact key press used, and then look it
up again with "control" tacked on if nothing was found.


git-svn-id: http://code.imagej.net/svn/imagej/trunk@4298 be23795f-a60e-40d4-ba19-055798872e57