Add initializer field to @Parameter annotation.
This allows a parameter to specify an initializer function that should
be called during the preprocessing chain, which will compute the
parameter's initial value.

The new InitPreprocessor plugin takes care of doing this, with priority
set *after* the ActiveDisplayPreprocessor and similar, but *before*
preprocessors of normal priority (so that they can count on the field
values already being initialized), as well as before the UI's input
harvester dialog is shown onscreen.

Note that initialized parameters are not marked as resolved; the
initializer merely computes the parameter's initial value, rather than
the final value desired for input to the plugin.

This used to be revision r4404.