Initialize only inputs, not outputs.
Firstly, it is probably very rare that an output would need an initial
value computed *before* the run() method gets called. The whole point of
run() is to populate output values based on input values, so
initializing outputs seems redundant. Secondly, for items flagged as
BOTH (input and output), the previous code would call the initializer
twice. For now, we'll initialize inputs only to avoid this.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4405 be23795f-a60e-40d4-ba19-055798872e57