Split Pivot UI and Pivot Application to 2 classes
This change allows the Pivot UI to work again, preventing two instances
of PivotUI from being needlessly launched.

The PivotUI launches the PivotApplication in a separate thread using the
ThreadService, so that it can access the necessary ImageJ services for
its context in its noargs constructor.

AFAICT, to instantiate a Pivot application one must use
DesktopApplicationContext.main, which invokes the given class name's
noargs constructor via reflection. So we must have a way to obtain
ImageJ services without constructor-based DI. So we use the fallback
mechanism of ImageJ.getContext() using the thread name.

This used to be revision r4429.