Simplify setInputs and setOutputs implementations
The PluginModule had a simple implementation that simply looped over the
map and called setInput/setOutput on each. This has now been propagated
into AbstractModule. As a side effect, the map is no longer cleared; in
the future we may want to provide an explicit mechanism for clearing
module values, but at the moment there isn't one.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4459 be23795f-a60e-40d4-ba19-055798872e57