Simplify setInputs and setOutputs implementations
The PluginModule had a simple implementation that simply looped over the
map and called setInput/setOutput on each. This has now been propagated
into AbstractModule. As a side effect, the map is no longer cleared; in
the future we may want to provide an explicit mechanism for clearing
module values, but at the moment there isn't one.

This used to be revision r4459.