Generalize default getInputs/getOutputs code
The PluginModule has a general mechanism for building the maps from the
input values, which has now been moved into AbstractModule. This
replaces the hardcoded unmodifiable maps wrapping AbstractModule's
internal map structures. The rationale for moving away from that is so
that subclasses of AbstractModules can override getInput, getOutput,
setInput and setOutput without needing to do so for getInputs or
getOutputs.

As a bonus, the returned map is now a mutable copy, in case client code
needs to make any modifications to it.

This used to be revision r4460.