Improve DynamicPlugin @Parameter tracking
This change causes DynamicPlugin @Parameter inputs and outputs to be
fully integrated into the Module API. Previously, all DynamicPlugin
module item values were stored in a hashtable (as per the DefaultModule
implementation). Now DynamicPlugin is smarter; it first checks whether a
given module item was declared via a @Parameter, and if so, it uses that
field to store and retrieve the item value, rather than the hashtable.

This used to be revision r4461.