Rename AxisAccelerator to AxisPositionHandler
The term "accelerator" refers to keyboard shortcut combinations shown in
the menu structure. So this always-active tool that handles dimensional
position changes is really an axis position handler that is mostly
unrelated to those accelerators in the menus.

Conversely, the class AcceleratorHandler is indeed responsible for
handling key presses that resemble, but do not match up exactly, with
the accelerators shown in the menu. In other words, AcceleratorHandler
does indeed handle accelerator key combinations for launching modules.

This used to be revision r4490.