Add support for global module initializer methods
This is in addition to per-item initializers. This global initializer is
a useful place to manipulate the number of module items (i.e., add or
remove items), whose individual initializers will be subsequently
invoked.

This change introduces a simple abstraction (ModuleRef) for calling
methods via reflection, to avoid repeating code. Currently it is used
for module initializers, item initializers, and item callbacks.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4516 be23795f-a60e-40d4-ba19-055798872e57