Add support for global module initializer methods
This is in addition to per-item initializers. This global initializer is
a useful place to manipulate the number of module items (i.e., add or
remove items), whose individual initializers will be subsequently
invoked.

This change introduces a simple abstraction (ModuleRef) for calling
methods via reflection, to avoid repeating code. Currently it is used
for module initializers, item initializers, and item callbacks.

This used to be revision r4516.