Change DataView getPosition and setPosition sigs
The methods now take an AxisType argument rather than an integer index.
This is more flexible, allowing position to be localized for any axis,
not just those of the associated Data object, in theory.

For now, the methods throw an exception if attempting to localize an
axis other than one of the associated Data object, but now that the API
change has been made, the behavior can be corrected in a later commit.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4542 be23795f-a60e-40d4-ba19-055798872e57