Redo how data object visibility is determined
The new method uses the data view's position, which can include
dimensional positions outside the linked data's N-space, in which case
those positions specify the value at which the data object should
appear. This is particularly useful for overlays, which are always 2D at
the moment.

Overlay visibility is still not fully fixed, however. More changes are
first needed regarding how overlays get created and linked to displays.

This used to be revision r4613.