Link JHotDraw Figure to OverlayView, not Overlay
This change removes the capability of Overlay objects to track their own
positions, and instead the OverlayView takes care of that (since
DataView is already LocalizableByAxis). The convention is that axis
positions outside the linked Data's N-space specify where along that
axis the view should be embedded. If the current display position in its
composite N-space does not include to the embedded axis position, the
view will be invisible. In this way, overlays should only appear over
the plane upon which they were originally drawn.

But it is not fully working yet; there are still bugs.

This used to be revision r4615.