Remove PluginFinder annotation functionality
This was a superfluous layer of discovery, which has been completely
replaced by the services framework. Now instead of LegacyPluginFinder,
the LegacyService takes care of adding its legacy plugins to the
PluginService's index, and similarly, the ScriptService can take care of
adding its scripts in the same way.

The problem with PluginFinders is that the PluginService would always
load all available PluginFinders on the classpath, and invoke them all,
which would cause problems for LegacyPluginFinder, since it assumed the
LegacyService was available. Now, the LegacyPluginFinder explicitly
needs a LegacyService to operate, and is only invoked *by* that
LegacyService, so all is well either way.

This used to be revision r4619.