Simplify application context creation & management
It is now possible to build an application context using preexisting
service instances. The logic for instantiating services has also been
improved to allow for more permissive matching of compatible services.

For example, previously if a service had a parameter of type
EventService, the framework would attempt to recursively instantiate
EventService.class. Now, if any existing service is assignable to an
EventService, it is used to fill the dependency.

This allows for a more interface-driven service design, where it is
easier to modify the core service implementations.

This commit also greatly simplifies the static tracking of application
contexts. The eventual goal is to eliminate static application context
tracking completely, but for now we track only a single static context
at a time, obtainable with the deprecated ImageJ.getContext() method.

This used to be revision r4646.