Remove redundant IService.initialize() method
Initialization is now done in the constructor, which is simpler and more
straightforward.

As a historical note, the reason initialize() was separate was so that
all services could first be instantiated, and then all initialized, in
order to limitedly handle circular dependencies. This mechanism was
important when PluginService and LegacyService were circularly dependent
due to the LegacyPluginFinder, but that issue is now resolved, and no
circular dependencies remain.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4647 be23795f-a60e-40d4-ba19-055798872e57