Move toplevel menu constants to separate class
This also declares constants for label, weight and mnemonic, and uses
them everywhere the relevant menus are referenced.

I tried to have MenuConstants define the @Menu annotations themselves,
but unfortunately the compiler does not understand references to Menu
instances from inside @Plugin annotation declarations.

This used to be revision r4690.