Add initial implementation of popup context menus
The idea is to support multiple menus through the use of "menu root" ID
strings. Each plugin can declare the menu root to which it belongs. So a
plugin designed to be accessed via a context menu can use a menu root
other than the default application menu (Plugin.APPLICATION_MENU_ROOT).

The UIService now provides a showContextMenu method for popping up the
context menu of a given menu root.

There is also an initial implementation of an always-active tool that
listens for popup menu trigger mouse events and displays the context
menu accordingly using the UIService.

This used to be revision r4705.