Updater: add the login() method to the AbstractUploader
A subclass of AbstractUploader can override this method using the
UserInterface class to implement an interactive login (but keep in
mind, this might run non-interactively, via cronjob or script, so
the login should always allow for presets).

While at it, remove some dead leftovers from Fiji Updater's login code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4754.