Updater (macosx): when installing the ImageJ launcher, patch Info.plist
Since Info.plist contains the information which executable is executed
and the ImageJ launcher obsoletes the Fiji launcher, and since
Info.plist is not in the Updater so people can change the default
VM arguments into it and it will never be overwritten by an update,
we have to patch that file to point to the new executable.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4777.