Updater: get rid of Util.prefix() altogether
For testing at least, it is important to be able to have
multiple instances of the Updater classes work on independent
ijRoots. Therefore, we cannot have a global prefix() function.

While at it, let prefix() -- which is now a non-static method of
FilesCollection -- return a File object since that is what you usually
need the prefixing for anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4802.