Updater: add API for moving things into place
To safeguard against partial overwrites as well as due to JRE's
inability to cope with .jar files that change while in use, we download
stuff into <imagejRoot>/update/. Sometimes it is quite handy to have
a Java API to move the files into their correct place (as opposed to
letting the ImageJ launcher handle it, requiring a restart).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4840.