Updater (JUnit tests): introduce a debug method to launch the Updater GUI
When things go awfully wrong, it is always nice to have a look at
things. Alas, when things go wrong in the Updater tests -- which are
fully headless -- one might wish for the Updater GUI to inspect the
state of affairs. Now one can.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r4842.