Remove superfluous "required = true" attributes
The required attribute is now true by default, which is a much more
common case. If you want to allow null values for Object parameters, you
can explicitly say "required = false" now instead. So far though, we
don't have a single example of that.

This used to be revision r4868.