TODO: Scripting: add a tag for compiled languages
This interface serves as a duct-typing-interface to label
ScriptEngineFactory implementations to tell ImageJ2 that the language
actually needs compilation prior to execution.

It might still require a companion interface (to be done in the future)
for ScriptEngine to add the compile() method returning a list of compile
products. But maybe it is enough to add that method to CompiledLanguage;
we will see when porting Refresh_Javas to ImageJ2 ;-)

Note: the ScriptEngineFactory must be labeled so that e.g. the Script
Editor can decide even without a ScriptEngine instance that there need to
be different labels in the Run menu (Compile and Run, as opposed to Run).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>