Add DatasetService, merge DatasetFactory methods
The static DatasetFactory class did not match the rest of the
application design, and provided no facility for tracking the
application context. With the DatasetService, it will be easier for Data
objects to belong to a specific context. It will also be easier to
override the behavior of dataset creation programmatically, if needed.

This used to be revision r4891.